comic-store-api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.22
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pymysql==1.1.0
alembic==1.12.1
python-dotenv==1.0.0